Add cache for metric objects
