+++ packages/opentelemetry/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsc --watch --preserveWatchOutput",
-    "test": "mocha --no-config --require ts-node/register src/index.test.ts"
+    "test": "mocha --no-config --require ts-node/register src/**/*.test.ts"
   },
   "dependencies": {
     "@grpc/grpc-js": "^1.8.8",
@@ -32,7 +32,11 @@
   },
   "devDependencies": {
     "@prairielearn/tsconfig": "*",
+    "@types/chai": "^4.3.4",
+    "@types/mocha": "^10.0.1",
     "@types/node": "^18.11.18",
+    "chai": "^4.3.7",
+    "mocha": "^10.2.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5"
   }
